perf: stash Obj headers in a 64bit int
The header was 128-bits long (`type`, `is_marked` and `next`), which could be
expressed in 49 bits only. So now they are stashed in a 64-bit int with the `is_marked` bit
in the first byte, the `next` pointer in the following six, and the `type` field in the least
significant byte.